remove rid.rs, and mv RID to tuple.rs
